% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.2nd.deriv.R
\name{F.2nd.deriv}
\alias{F.2nd.deriv}
\title{F.2nd.deriv}
\usage{
F.2nd.deriv(x, FUN, ...)
}
\arguments{
\item{x}{A vector}

\item{FUN}{function to be inputted}
}
\value{

}
\description{
This function calculates the second derivative (hessian value)
for a function
}
