% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModMatrix.R
\name{getModMatrix}
\alias{getModMatrix}
\title{getModMatrix}
\usage{
getModMatrix(f, ncells)
}
\arguments{
\item{f}{model formula}

\item{ncells}{number of cells in the study area}
}
\description{
Mainly a helper function, this creates the model matrix for
the optimization process. Not recommended for isolated use.
}
