% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSSFLogLike.R
\name{iSSFLogLike}
\alias{iSSFLogLike}
\title{iSSFLogLike}
\usage{
iSSFLogLike(beta, X1, locations, k1)
}
\arguments{
\item{beta}{Coefficients to be estimated}

\item{X1}{Data from the covariates for the iSSF component}

\item{locations}{data frame of GPS fixes}

\item{k1}{Number of covariates for the iSSF}
}
\description{
The log-likelihood formulation for the iSSF.
Not recommended for isolated use.
}
