% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixCalcFastTwo.R
\name{matrixCalcFastTwo}
\alias{matrixCalcFastTwo}
\title{matrixCalcFastTwo}
\usage{
matrixCalcFastTwo(matrixOne, multTimes, matrixOfDiag)
}
\arguments{
\item{matrixOne}{Output to be multiplied multTimes times}

\item{multTimes}{number of matrix multiplications to perform}

\item{matrixOfDiag}{Empty matrix of the form diag(1, ncells)}
}
\description{
Calls the rcpp file. Not recommended for isolated use.
}
