% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdissf.R
\name{pdissf}
\alias{pdissf}
\title{pdissf}
\usage{
pdissf(
  habitatDF,
  CellID,
  maxLagArg,
  habitatCellID = "unitID",
  iSSFCovars = NULL,
  probDetCovars = NULL,
  distColumns = NULL
)
}
\arguments{
\item{habitatDF}{Dataframe of habitat information}

\item{CellID}{Column that contains the unique ID for a cell location in the study
area}

\item{maxLagArg}{Number of matrix multiplications to toggle the matrix by
squaring algorithm. The default is 4 and it is not recommended to change this.}

\item{habitatCellID}{Column in the habitat DF that contains unique ID for the
habitat data}

\item{iSSFCovars}{covariates in the integrated step selection function model}

\item{probDetCovars}{covariates for the probability of detection}

\item{distColumns}{If distance is needed as a covariate, in either iSSF or the
probability of detection, then identify the columns that contain this information}
}
\description{
Helper function to set up the log-likelihood estimation
}
