% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdissf.R
\name{pdissf}
\alias{pdissf}
\title{pdissf}
\usage{
pdissf(
  habitatDF,
  CellID,
  habitatCellID,
  maximumGap = 3,
  iSSFCovars = NULL,
  probDetCovars = NULL,
  distColumns = NULL
)
}
\arguments{
\item{habitatDF}{Dataframe of habitat information}

\item{CellID}{Column that contains the unique ID for a cell location in the study
area}

\item{habitatCellID}{Column in the habitat DF that contains unique ID for the
habitat data}

\item{maximumGap}{Maximum allowable number of consecutive missing fixes. Default is 3
(3 consecutive missing locations). Large gaps will cause an exponential increase in
memory requirements and computing time. This may result in convergence issues.}

\item{iSSFCovars}{covariates in the integrated step selection function model}

\item{probDetCovars}{covariates for the probability of detection}

\item{distColumns}{If distance is needed as a covariate, in either iSSF or the
probability of detection, then identify the columns that contain this information}
}
\description{
Helper function to set up the log-likelihood estimation
}
\examples{
library(PDiSSF)
data(habitat)
data(locations)

# Fix success rate
mean(!is.na(locations$unitID))

# Computing time for larger data sets may vary WIDEL

# integrated step selection function (iSSF) using step length,
pdissf(habitatDF = habitat,
       habitatCellID = 'unitID',
       CellID = locations$unitID, 
       iSSFCovars = c('distance', 'prctSage', 'elevation'), 
       probDetCovars = NULL, 
       distColumns = c('utmX','utmY'))

# PDRSF including step length
pdissf(habitatDF = habitat, 
       habitatCellID = 'unitID',
       CellID = locations$unitID, 
       iSSFCovars = c('distance', 'prctSage', 'elevation'), 
       probDetCovars = 'prctSage', distColumns = c('utmX','utmY'),
       maximumGap = 4)


}
