% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDiSSFModel.R
\name{PDiSSFModel}
\alias{PDiSSFModel}
\title{PDiSSFModel}
\usage{
PDiSSFModel(selection, p, locations, ncells, maximumGap, iSSFCovar, LogCovar)
}
\arguments{
\item{selection}{Model for habitat selection}

\item{p}{Model for encounter probabilities}

\item{locations}{Dataframe of animal locations}

\item{ncells}{Number of cells in the study area}

\item{maximumGap}{Maximum allowable number of consecutive missing fixes. Default is 4
(3 consecutive missing locations). Larger max lags will increase computing time
and may result in non convergence.}

\item{iSSFCovar}{Covariates for the iSSF}

\item{LogCovar}{Covariate for ...}
}
\description{
Sets up the parameter estimation for the log-likelihood function
}
