% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDiSSFModel.R
\name{PDiSSFModel}
\alias{PDiSSFModel}
\title{PDiSSFModel}
\usage{
PDiSSFModel(selection, p, locations, ncells, maxLagArg, iSSFCovar, LogCovar)
}
\arguments{
\item{selection}{Model for habitat selection}

\item{p}{Model for encounter probabilities}

\item{locations}{Dataframe of animal locations}

\item{ncells}{Number of cells in the study area}

\item{maxLagArg}{Number of matrix multiplications to toggle the matrix by
squaring algorithm. The default is 4 and it is not recommended to change this.}

\item{iSSFCovar}{Covariates for the iSSF}

\item{LogCovar}{Covariate for ...}
}
\description{
Sets up the parameter estimation for the log-likelihood function
}
